+++ package.json
@@ -13,8 +13,11 @@
     "postversion": "git push; git push --tags; npm publish"
   },
   "dependencies": {
+    "fuzzy": "^0.1.3",
     "glob": "^7.1.2",
-    "plop": "^1.9.0"
+    "inquirer-autocomplete-prompt": "^0.11.1",
+    "plop": "^1.9.0",
+    "promise": "^8.0.1"
   },
   "devDependencies": {
     "jest": "^21.2.1",
